Fix aesthetic errors
We had some issue with python complaining that at the end of the script, some things were already deleted. This didn't affect the codes output, but was aestheticly undesirable, and rooted in a way to work around some ROOT things going out of scope. I implemented now a new way to keep the things in scope.